Change type of literal_length to int to avoid warnings

+++ postgis/lwgeom_in_marc21.c
@@ -185,7 +185,7 @@ static double parse_geo_literal(char *literal) {
 	char *dec;
 	char *csl;
 
-	size_t literal_length;
+	int literal_length;
 
 	char hemisphere_sign = literal[0];
 	int start_literal = 0;
